[Model Runner V2] Fuse probabilistic rejection sample kernels
